refactor: move application start logic into bot file
Eventually, the main bot class will be housed in a 'core' directory. This will set us up better for the restructure
